changed a few ints to size_t when dealing with kmers on a rank

+++ src/kcount/kcount_cpu.cpp
@@ -75,7 +75,7 @@ void SeqBlockInserter<MAX_K>::process_seq(string &seq, kmer_count_t depth, dist_
   if (!depth) depth = 1;
   auto kmer_len = Kmer<MAX_K>::get_k();
   Kmer<MAX_K>::get_kmers(kmer_len, seq, state->kmers);
-  for (int i = 0; i < state->kmers.size(); i++) {
+  for (size_t i = 0; i < state->kmers.size(); i++) {
     state->bytes_kmers_sent += sizeof(KmerAndExt<MAX_K>);
     Kmer<MAX_K> kmer_rc = state->kmers[i].revcomp();
     if (kmer_rc < state->kmers[i]) state->kmers[i] = kmer_rc;
@@ -212,7 +212,7 @@ class KmerMapExts {
   size_t sum_probe_lens = 0;
   size_t max_probe_len = 0;
   vector<KmerExtsCounts> counts;
-  int iter_pos = 0;
+  size_t iter_pos = 0;
   const int N_LONGS = Kmer<MAX_K>::get_N_LONGS();
   const uint64_t KEY_EMPTY = 0xffffffffffffffff;
 
@@ -229,10 +229,10 @@ class KmerMapExts {
   }
 
   pair<KmerExtsCounts *, bool> insert(const Kmer<MAX_K> &kmer, bool override_singletons) {
-    int slot = kmer.hash() % capacity;
-    int start_slot = slot;
-    const int MAX_PROBE = (capacity < KCOUNT_HT_MAX_PROBE ? capacity : KCOUNT_HT_MAX_PROBE);
-    for (int i = 1; i <= MAX_PROBE; i++) {
+    size_t slot = kmer.hash() % capacity;
+    size_t start_slot = slot;
+    const size_t MAX_PROBE = (capacity < KCOUNT_HT_MAX_PROBE ? capacity : KCOUNT_HT_MAX_PROBE);
+    for (size_t i = 1; i <= MAX_PROBE; i++) {
       if (keys[slot].get_longs()[N_LONGS - 1] == KEY_EMPTY) {
         keys[slot] = kmer;
         sum_probe_lens += i;
@@ -250,7 +250,7 @@ class KmerMapExts {
     if (override_singletons) {
       // reset variables for search
       slot = start_slot;
-      for (int i = 1; i <= MAX_PROBE; i++) {
+      for (size_t i = 1; i <= MAX_PROBE; i++) {
         assert(kmer != keys[slot]); // FIXME? probe_lens[slot] != 0
         if (counts[slot].count == 1) {
           num_singleton_overrides++;
@@ -423,7 +423,7 @@ HashTableInserter<MAX_K>::~HashTableInserter() {
 }
 
 template <int MAX_K>
-void HashTableInserter<MAX_K>::init(int num_elems, bool use_qf) {
+void HashTableInserter<MAX_K>::init(size_t num_elems, bool use_qf) {
   state = new HashTableInserterState();
   state->using_ctg_kmers = false;
   double free_mem = get_free_mem();
@@ -443,7 +443,7 @@ void HashTableInserter<MAX_K>::init(int num_elems, bool use_qf) {
 }
 
 template <int MAX_K>
-void HashTableInserter<MAX_K>::init_ctg_kmers(int max_elems) {
+void HashTableInserter<MAX_K>::init_ctg_kmers(size_t max_elems) {
   state->using_ctg_kmers = true;
 }
